Changes in package structure
Changed the package structure of the project
