+++ src/com/cactus/BME/PageObjectModel/CrmLoginPage.java
@@ -1,11 +1,10 @@
-package PageObjectModel;
+package com.cactus.BME.PageObjectModel;
 
-import org.openqa.selenium.By;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class CrmLoginPage {
